Resolve duplicate 'useFeature' issue
