--- libs/ui-lib/lib/common/features/index.ts
@@ -1,1 +1,0 @@
-export * from './featureGate';
