+++ libs/ui-lib/lib/common/index.ts
@@ -3,7 +3,6 @@ export * from './api';
 export * from './types';
 export * from './config';
 export * from './configurations';
-export * from './features';
 export * from './reducers';
 export * from './selectors';
 export * from './hooks';
