+++ libs/ui-lib/lib/ocm/hooks/use-feature.ts
@@ -1,6 +1,6 @@
-import { AssistedInstallerFeatureType } from '../../common';
 import { useSelectorDay1 } from '../store';
 import { isFeatureEnabled } from '../store/slices/feature-flags/selectors';
+import { AssistedInstallerFeatureType } from './types';
 
 export const useFeature = (feature: AssistedInstallerFeatureType): boolean => {
   return useSelectorDay1(isFeatureEnabled(feature));
