+++ libs/ui-lib/lib/ocm/store/slices/feature-flags/selectors.ts
@@ -1,7 +1,7 @@
-import type { FeatureListType } from '../../../../common/features/featureGate';
-import type { RootStateDay1 } from '../../store-day1';
 import { createSelector } from '@reduxjs/toolkit';
 import type { Selector } from 'reselect';
+import type { RootStateDay1 } from '../../store-day1';
+import { FeatureListType } from '../../../hooks/types';
 
 export const selectFeatureFlagsSlice = (state: RootStateDay1) => state.featureFlags;
